Dos features:

- BIR-19: la nota de una birra admite un decimal (3,8), y el piso baja de
  1 a 0 porque con carga numérica un 0 es un voto real, distinto de no
  haber votado. beer_ratings.rating pasa a numeric(2,1) (V10).

- BIR-20: analíticas en el dashboard de admin. Vista v_contributions
  (V11) como única definición de qué es un aporte; tabla traffic_sessions
  (V12/V13) que cuenta visitantes sin guardar IP ni user agent; seis
  métricas en AnalyticsRepo detrás de un endpoint de moderador; tres
  primitivas SVG propias (los seis gráficos costaron 2,2 kb gzip, no una
  librería); y el dashboard a ancho completo en desktop.

69 tests de backend en verde. Web y app compilan.

OJO con el orden de deploy: los Android v0.5.0 instalados decodifican
rating como Int y pierden los votos propios en silencio al ver un
decimal. Backend primero, después el APK 0.6.0, y recién ahí la web.

Y esto empieza a recolectar datos de visitantes: hay que declararlo en
BIR-15 y BIR-16 antes de la próxima submission a tienda.

+++ app/src/main/java/com/birrapp/data/model/Models.kt
@@ -122,14 +122,16 @@ data class BarDetail(
     val body: String? = null,
     val ageDays: Int,
     val mine: Boolean = false,
-    val rating: Int? = null,
+    /** De 0 a 5, con un decimal: alguien pudo cargar un 3,8 desde la web. */
+    val rating: Double? = null,
 )
 
 /** Lo que votó el usuario en un bar, por birra. */
 @Serializable data class MyRating(
     val styleSlug: String,
     val brandSlug: String? = null,
-    val rating: Int,
+    /** De 0 a 5, con un decimal. */
+    val rating: Double,
 )
 
 @Serializable data class NewRatingRequest(
